requirements.txt
Adafruit-Blinka==6.9.2
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-framebuf==1.4.7
adafruit-circuitpython-ssd1306==2.11.4
Adafruit-PlatformDetect==3.13.1
Adafruit-PureIO==1.1.8
certifi==2020.12.5
chardet==4.0.0
idna==2.10
libusb1==1.9.2
ndeflib==0.3.3
nfcpy==1.0.3
pkg-resources==0.0.0
pyDes==2.0.1
pyftdi==0.53.1
pyserial==3.5
pyusb==1.1.1
requests==2.25.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
urllib3==1.26.4
